news-dashboard/requirements.txt
flask>=3.0.3
psycopg2-binary>=2.8.4
python-dotenv>=1.0.1
